services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "langfuse>=3.12.1",
    "llm-dataclass>=1.1.5",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sse-starlette>=3.1.2",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["api","../../core"]
include = ["*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
